Main copy algorithm improved. New way to manage file datetime provided.

+++ qthsynchronize.cpp
@@ -23,8 +23,6 @@
 QThSynchronize::QThSynchronize(Profile profile, bool Simulation) : QThread() {
     this->profile= profile;
     this->Simulation= Simulation;
-    DoStart= true;
-    FilesCopied= DirecoriesCopied= FilesDeleted= DirectoriesDeleted= 0;
 }
 
 QThSynchronize::~QThSynchronize() {
@@ -62,91 +60,72 @@ void QThSynchronize::DirectoryCheckToCopy(QString Source, QString Destination, Q
     QStringList QSLFiles= QDPathSource.entryList(QDir::AllEntries | QDir::NoDotAndDotDot | QDir::Hidden);
     for (int count= 0; count< QSLFiles.length(); count++) {
         QFileInfo QFIPathSource(QDPathSource.absolutePath()+ "/"+ QSLFiles.at(count));
-        bool Finded= false;
+        bool CanGo= true;
         for (int count_2= 0; count_2< profile.QSLExclusionFiles.length(); count_2++) {
             if (QString(QDPathSource.absolutePath()+ "/"+ QSLFiles.at(count)).compare(profile.QSLExclusionFiles.at(count_2), Qt::CaseInsensitive)== 0) {
-                Finded= true;
+                CanGo= false;
                 break;
             }
         }
-        if (!Finded) {
+        if (CanGo) {
             QFileInfo QFIPathDestination(Destination+ "/"+ QFIPathSource.absoluteFilePath().right(QFIPathSource.absoluteFilePath().length()- Source.length()));
             if (DoStart && QFIPathSource.isDir()) {
-                if (!QFIPathSource.isSymLink() &! QFIPathSource.isBundle()) {
+                if (!QFIPathSource.isSymLink() && !QFIPathSource.isBundle()) {
                     if (!QFIPathDestination.exists()) {
                         if (Simulation) {
                             DirecoriesCopied++;
                             emit OnGenericEvent(EVENT_TYPE_JOB_STATUS, FilesCopied, DirecoriesCopied, FilesDeleted, DirectoriesDeleted, "");
-                            emit OnLog(LOG_TYPE_STANDARD, QFIPathSource.absoluteFilePath()+ tr(" created"));
+                            emit OnLog(LOG_TYPE_STANDARD, "["+ QFIPathSource.absoluteFilePath()+ "]"+ tr(" created"));
                         } else {
                             QDir QDPathDestination;
-                            if (!QDPathDestination.mkpath(QFIPathDestination.absoluteFilePath())) emit OnLog(LOG_TYPE_ERROR, tr("Error on create folder ")+ QFIPathDestination.absoluteFilePath());
+                            if (!QDPathDestination.mkpath(QFIPathDestination.absoluteFilePath())) emit OnLog(LOG_TYPE_ERROR, tr("Error on create folder ")+ "["+ QFIPathDestination.absoluteFilePath()+ "]");
                             else {
-                                if (!SetFileDateTime(QFIPathDestination.absoluteFilePath(), QFIPathSource.birthTime(), QFIPathSource.lastRead(), QFIPathSource.lastModified())) emit OnLog(LOG_TYPE_ERROR, tr("Error on change folder date time ")+ QFIPathDestination.absoluteFilePath());
+                                if (!SetFileDateTime(QFIPathDestination.absoluteFilePath(), QFIPathSource.birthTime(), QFIPathSource.lastRead(), QFIPathSource.lastModified()))
+                                    emit OnLog(LOG_TYPE_ERROR, tr("Error on change folder date time ")+ "["+ QFIPathDestination.absoluteFilePath()+ "]");
                                 DirecoriesCopied++;
                                 emit OnGenericEvent(EVENT_TYPE_JOB_STATUS, FilesCopied, DirecoriesCopied, FilesDeleted, DirectoriesDeleted, "");
-                                emit OnLog(LOG_TYPE_STANDARD, QFIPathSource.absoluteFilePath()+ tr(" created"));
+                                emit OnLog(LOG_TYPE_STANDARD, "["+ QFIPathSource.absoluteFilePath()+ "]"+ tr(" created"));
                             }
                         }
                     }
                     DirectoryCheckToCopy(Source, Destination, QFIPathSource.absoluteFilePath(), IgnoreFileDate, OnlyIfBest);
                 }
             } else if (DoStart && QFIPathSource.isFile()) {
-                bool CanGo= false;
-                if (!QFIPathDestination.exists()) CanGo= true;
-                else {
+                QString Reason;
+                CanGo= false;
+                if (!QFIPathDestination.exists()) {
+                    CanGo= true;
+                    Reason= tr("(not exists)");
+                } else {
                     if (!OnlyIfBest) {
-                        if (QFIPathSource.size()!= QFIPathDestination.size() || QFIPathSource.lastModified().toString("yyyyMMddhhmmss").compare(QFIPathDestination.lastModified().toString("yyyyMMddhhmmss"))!= 0) CanGo= true;
-                        else {
-                            #ifdef Q_OS_WIN
-                                if (QFIPathSource.created().toString("yyyyMMddhhmmss").compare(QFIPathDestination.created().toString("yyyyMMddhhmmss"))!= 0) CanGo= true;
-                            #endif
+                        if (QFIPathSource.size()!= QFIPathDestination.size()
+                                || QFIPathSource.lastModified().toString("yyyyMMddhhmmss").compare(QFIPathDestination.lastModified().toString("yyyyMMddhhmmss"))!= 0) {
+                            CanGo= true;
+                            Reason= tr("(file date time not equals)");
                         }
+                        /*else {
+                            #ifdef Q_OS_WIN
+                                if (QFIPathSource.birthTime().toString("yyyyMMddhhmmss").compare(QFIPathDestination.birthTime().toString("yyyyMMddhhmmss"))!= 0) CanGo= true;
+                            #endif
+                        }*/
                     } else {
                         if (QFIPathSource.lastModified().toString("yyyyMMddhhmmss").compare(QFIPathDestination.lastModified().toString("yyyyMMddhhmmss"))> 0) {
                             CanGo= true;
-                            qDebug() << QFIPathSource.lastModified().toString("yyyyMMddhhmmss") << QFIPathDestination.lastModified().toString("yyyyMMddhhmmss");
+                            Reason= tr("(source file date time greater than destination)");
                         }
-                        else {
+                        /*else {
                             #ifdef Q_OS_WIN
-                                if (QFIPathSource.created().toString("yyyyMMddhhmmss").compare(QFIPathDestination.created().toString("yyyyMMddhhmmss"))> 0) CanGo= true;
+                                if (QFIPathSource.birthTime().toString("yyyyMMddhhmmss").compare(QFIPathDestination.birthTime().toString("yyyyMMddhhmmss"))> 0) CanGo= true;
                             #endif
-                        }
+                        }*/
                     }
                 }
                 if (CanGo) {
                     if (Simulation) {
                         FilesCopied++;
                         emit OnGenericEvent(EVENT_TYPE_JOB_STATUS, FilesCopied, DirecoriesCopied, FilesDeleted, DirectoriesDeleted, "");
-                        emit OnLog(LOG_TYPE_STANDARD, QFIPathSource.absoluteFilePath()+ tr(" copied"));
-                    } else {
-                        QFile QFSource(QFIPathSource.absoluteFilePath());
-                        QFile QFDestination(QFIPathDestination.absoluteFilePath());
-                        if (QFSource.open(QIODevice::ReadOnly)) {
-                            if (QFDestination.open(QIODevice::WriteOnly)) {
-                                qint64 Position= 0;
-                                emit OnGenericEvent(EVENT_TYPE_FILE_STATUS, 0, 0, 0, 0, "");
-                                emit OnGenericEvent(EVENT_TYPE_FILE, 0, 0, 0, 0, QFSource.fileName());
-                                QDateTime QDTLastEvent= QDateTime::currentDateTime();
-                                while(DoStart && !QFSource.atEnd()) {
-                                    Position+= 4096;
-                                    if (QDTLastEvent.msecsTo(QDateTime::currentDateTime())> 250) {
-                                        emit OnGenericEvent(EVENT_TYPE_FILE_STATUS, static_cast<int>((Position * 100) / QFSource.size()), 0, 0, 0, "");
-                                        QDTLastEvent= QDateTime::currentDateTime();
-                                    }
-                                    QFDestination.write(QFSource.read(4096));
-                                    QFDestination.flush();
-                                }
-                                emit OnGenericEvent(EVENT_TYPE_FILE_STATUS, 100, 0, 0, 0, "");
-                                QFDestination.close();
-                                FilesCopied++;
-                                emit OnGenericEvent(EVENT_TYPE_JOB_STATUS, FilesCopied, DirecoriesCopied, FilesDeleted, DirectoriesDeleted, "");
-                                if (!SetFileDateTime(QFIPathDestination.absoluteFilePath(), QFIPathSource.birthTime(), QFIPathSource.lastRead(), QFIPathSource.lastModified())) emit OnLog(LOG_TYPE_ERROR, QFDestination.fileName()+ tr(" error on change file date time!"));
-                                else emit OnLog(LOG_TYPE_STANDARD, QFSource.fileName()+ tr(" copied"));
-                            } else OnLog(LOG_TYPE_ERROR, QFSource.fileName()+ tr(" error on copy to ")+ QFDestination.fileName());
-                            QFSource.close();
-                        } else OnLog(LOG_TYPE_ERROR, QFSource.fileName()+ tr(" error on copy to ")+ QFDestination.fileName());
-                    }
+                        emit OnLog(LOG_TYPE_STANDARD, "["+ QFIPathSource.absoluteFilePath()+ "]"+ tr(" copied")+ " "+ Reason);
+                    } else FileCopy(QFIPathSource, QFIPathDestination);
                 }
             }
         }
@@ -161,14 +140,14 @@ void QThSynchronize::DirectoryCheckToDelete(QString Source, QString Destination,
         QFileInfo QFIPathSource(QDPathSource.absolutePath()+ "/"+ QSLFiles.at(count));
         QFileInfo QFIPathDestination(Destination+ "/"+ QFIPathSource.absoluteFilePath().right(QFIPathSource.absoluteFilePath().length()- Source.length()));
         if (DoStart && QFIPathSource.isDir()) {
-            if (!QFIPathSource.isSymLink() &! QFIPathSource.isBundle()) {
+            if (!QFIPathSource.isSymLink() && !QFIPathSource.isBundle()) {
                 if (!QFIPathDestination.exists()) {
                     QDir QDPathSource(QFIPathSource.absoluteFilePath());
                     if (Simulation || QDPathSource.removeRecursively()) {
                         DirectoriesDeleted++;
                         emit OnGenericEvent(EVENT_TYPE_JOB_STATUS, FilesCopied, DirecoriesCopied, FilesDeleted, DirectoriesDeleted, "");
-                        emit OnLog(LOG_TYPE_STANDARD, QFIPathSource.absoluteFilePath()+ tr(" deleted"));
-                    } else emit OnLog(LOG_TYPE_ERROR, QFIPathSource.absoluteFilePath()+ tr(" error on delete!"));
+                        emit OnLog(LOG_TYPE_STANDARD, "["+ QFIPathSource.absoluteFilePath()+ "]"+ tr(" deleted"));
+                    } else emit OnLog(LOG_TYPE_ERROR, "["+ QFIPathSource.absoluteFilePath()+ "]"+ tr(" error on delete!"));
                 } else DirectoryCheckToDelete(Source, Destination, QFIPathSource.absoluteFilePath());
             }
         } else if (DoStart && QFIPathSource.isFile()) {
@@ -177,9 +156,51 @@ void QThSynchronize::DirectoryCheckToDelete(QString Source, QString Destination,
                 if (Simulation || QFSource.remove()) {
                     FilesDeleted++;
                     emit OnGenericEvent(EVENT_TYPE_JOB_STATUS, FilesCopied, DirecoriesCopied, FilesDeleted, DirectoriesDeleted, "");
-                    emit OnLog(LOG_TYPE_STANDARD, QFIPathSource.absoluteFilePath()+ tr(" deleted"));
-                } else emit OnLog(LOG_TYPE_ERROR, QFIPathSource.absoluteFilePath()+ tr(" error on delete!"));
+                    emit OnLog(LOG_TYPE_STANDARD, "["+ QFIPathSource.absoluteFilePath()+ "]"+ tr(" deleted"));
+                } else emit OnLog(LOG_TYPE_ERROR, "["+ QFIPathSource.absoluteFilePath()+ "]"+ tr(" error on delete!"));
             }
         }
     }
+}
+
+bool QThSynchronize::FileCopy(QFileInfo &QFIPathSource, QFileInfo &QFIPathDestination) {
+    bool Result= false;
+    QFile QFSource(QFIPathSource.absoluteFilePath());
+    QFile QFDestination(QFIPathDestination.absoluteFilePath());
+    if (QFSource.open(QIODevice::ReadOnly)) {
+        if (QFDestination.open(QIODevice::WriteOnly)) {
+            qint64 Position= 0;
+            emit OnGenericEvent(EVENT_TYPE_FILE_STATUS, 0, 0, 0, 0, "");
+            emit OnGenericEvent(EVENT_TYPE_FILE, 0, 0, 0, 0, QFSource.fileName());
+            QDateTime QDTLastEvent= QDateTime::currentDateTime();
+            while (DoStart && !QFSource.atEnd()) {
+                Position+= 4096;
+                if (QDTLastEvent.msecsTo(QDateTime::currentDateTime())> 250) {
+                    emit OnGenericEvent(EVENT_TYPE_FILE_STATUS, static_cast<int>((Position * 100) / QFSource.size()), 0, 0, 0, "");
+                    QDTLastEvent= QDateTime::currentDateTime();
+                }
+                QByteArray QBAByteIn= QFSource.read(1024 * 128);
+                if (QBAByteIn.length()> 0) {
+                    QFDestination.flush();
+                    if (QFDestination.write(QBAByteIn)!= QBAByteIn.size()) {
+                        emit OnLog(LOG_TYPE_ERROR, "["+ QFSource.fileName()+ "]"+ tr(" unable to copy file"));
+                        break;
+                    }
+                }
+            }
+            emit OnGenericEvent(EVENT_TYPE_FILE_STATUS, 100, 0, 0, 0, "");
+            QFDestination.close();
+            if (QFSource.size()== QFDestination.size()) {
+                FilesCopied++;
+                emit OnGenericEvent(EVENT_TYPE_JOB_STATUS, FilesCopied, DirecoriesCopied, FilesDeleted, DirectoriesDeleted, "");
+                if (!SetFileDateTime(QFIPathDestination.absoluteFilePath(), QFIPathSource.birthTime(), QFIPathSource.lastRead(), QFIPathSource.lastModified())) emit OnLog(LOG_TYPE_ERROR, "["+ QFDestination.fileName()+ "]"+ tr(" error on change file date time!"));
+                else {
+                    emit OnLog(LOG_TYPE_STANDARD, "["+ QFSource.fileName()+ "]"+ tr(" copied"));
+                    Result= true;
+                }
+            } else emit OnLog(LOG_TYPE_ERROR, "["+ QFSource.fileName()+ "]"+ tr(" error on copy to ")+ QFDestination.fileName());
+        } else emit OnLog(LOG_TYPE_ERROR, "["+ QFSource.fileName()+ "]"+ tr(" error on copy to ")+ QFDestination.fileName());
+        QFSource.close();
+    } else emit OnLog(LOG_TYPE_ERROR, "["+ QFSource.fileName()+ "]"+ tr(" error on copy to ")+ QFDestination.fileName());
+    return Result;
 }
